requirements.txt
transformers
torch
pandas
numpy
Pillow
tabulate